see #18864 - TaggingPreset: use short type for numeric field
Reduces retained size of TaggingPreset from 1_758_840 to 1_750_272 (de).

git-svn-id: https://josm.openstreetmap.de/svn/trunk@16042 0c6e7542-c601-0410-84e7-c038aed88b3b

+++ src/org/openstreetmap/josm/gui/tagging/presets/items/Roles.java
@@ -55,7 +55,7 @@ public class Roles extends TaggingPresetItem {
         /** Is this role required at least once in the relation? */
         public boolean required; // NOSONAR
         /** How often must the element appear */
-        private long count;
+        private short count;
 
         public void setType(String types) throws SAXException {
             this.types = getType(types);
@@ -88,7 +88,7 @@ public class Roles extends TaggingPresetItem {
         }
 
         public void setCount(String count) {
-            this.count = Long.parseLong(count);
+            this.count = Short.parseShort(count);
         }
 
         /**
